update hikyuu script

+++ python/hikyuu/00_hikyuu.sh
@@ -25,14 +25,18 @@ then
   echo "found boost devel "
   echo "/usr/include" > boostroot.txt
 else
-  rm boostroot.txt
-  boostver=78 # for boost 1.74.0
+  [ -f "boostroot.txt" ] && rm boostroot.txt
+  # boostver=78 # for boost 1.74.0
+  . ./hikyuuEnv.sh
   boostfile="boost_1_${boostver}_0.tar.gz"
   # [ ! -f ${boostfile} ] && wget -c -O $boostfile https://dl.bintray.com/boostorg/release/1.${boostver}.0/source/boost_1_${boostver}_0.tar.gz
   [ ! -f ${boostfile} ] && wget -c -O $boostfile https://boostorg.jfrog.io/artifactory/main/release/1.${boostver}.0/source/boost_1_${boostver}_0.tar.gz
   # 第一次需要执行bootstrap,sh 产生配置文件
-:'i如果使用 Anaconda 的 python，需手工修改 boost 根目录下的 project-config.jam 文件, 找到 “using python” 所在行，手工添加python的版本、可执行文件、include目录，如：using python : 3.7 : “/Users/ljh/opt/anaconda3/bin/python3.7” : /Users/ljh/opt/anaconda3/include/python3.7m ;'
-  [ ! -d "boost_1_${boostver}_0" ] && tar xzvf ${boostfile} && cd "boost_1_${boostver}_0" && ./bootstrap.sh &
+: <<'EOF'
+如果使用Anaconda自带的python，在boost源码目录中执行： ./bootstrap.sh,生成project=-config.jam文件；
+需手工修改boost根目录下的 project-config.jam 文件, 找到 “using python” 所在行，手工添加python的版本、可执行文件、include目录，如：using python : 3.7 : “/Users/ljh/opt/anaconda3/bin/python3.7” : /Users/ljh/opt/anaconda3/include/python3.7m ;
+EOF
+  [ ! -d "boost_1_${boostver}_0" ] && tar xzvf ${boostfile} && cd "boost_1_${boostver}_0" && ./bootstrap.sh --exec-prefix=/usr/local &
 fi
 # 要改boost目录下的配置，指定python版本和路径
 
@@ -43,7 +47,9 @@ if [ ! -d "hikyuu" ]
 then
   git clone https://github.com/fasiondog/hikyuu.git --recursive --depth 1
 else
-  cd hikyuu && git pull && [ -f "../hikyuu.patch" ] && git apply ../hikyuu.patch ; cd ..
+  # cd hikyuu && git pull && [ -f "../hikyuu.patch" ] && git apply ../hikyuu.patch ; cd ..
+  cd hikyuu && git pull
+  cd ..
 fi
 
 # 自己fork的版本
